tools/sdk_sample_converter/requirements.txt
azure-identity
openai
python-dotenv
PyGithub
